Python/http/image/requirements.txt
cycler==0.11.0
dnspython==2.2.1
fonttools==4.33.3
imutils==0.5.4
kiwisolver==1.4.3
matplotlib==3.5.2
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.1.1
pymongo==4.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
